Compiling with PETSc 3.6.0 with debugging enabled.
This uncovered several places where we were closing PETSc Vecs
unnecessarily, which is an error when debugging is enabled.

[0]PETSC ERROR: Object is in wrong state
[0]PETSC ERROR:  Vec is locked read only, argument # 1

It seems safe to assume that the Vecs we are passed in the residual
and jacobian callback functions are closed, and therefore if we don't
call VecSetValues on them, we shouldn't attempt to close them
ourselves.

Refs #604.
Refs idaholab/moose#5235.