Remove enforcement of constraints from __libmesh_petsc_snes_{residual,jacobian}.
This is required to get libmesh working with debugging enabled in
PETSc: the solution vector is locked read only during residual
computation, so we are not allowed to modify it there.  We can't just
remove this code, however, we do need to enforce the constraints
somewhere.  The current idea is to possibly do that during the
application of "dampers" instead.